Feat - Add ability to delete uploaded files

+++ client/public/modules/delete-file.js
@@ -1,15 +1,29 @@
 // Delete the respective file when button is clicked
 export default function deleteFile() {
-  const deleteButton = document.querySelector('.delete-button');
+  document.addEventListener('click', async (event) => {
+    if (event.target.classList.contains('delete-button')) {
+      const deleteButton = event.target;
+      const fileContainer = event.target.parentElement;
+      const fileName = event.target.previousElementSibling.textContent;
 
-  deleteButton.addEventListener('click', async () => {
-    try {
-      // Send DELETE request to the specified endpoint
-      await fetch('http://localhost:3000/deleteFile', {
-      method: 'DELETE',
-    });
-    } catch (error) {
-      console.error('Error:', error.message);
+      try {
+        // Send DELETE request to the specified endpoint
+        const response = await fetch(`http://localhost:3000/deleteFile/${fileName}`, {
+            method: 'DELETE',
+          }
+        );
+        
+        deleteButton.remove();
+        fileContainer.remove();
+
+        if (response.ok) {
+          console.log(await response.json());
+        } else {
+          console.error(await response.json());
+        }
+      } catch (error) {
+        console.error('Error:', error.message);
+      }
     }
   });
 }
